archive/PoliQ_App (Error Resolving in Agents)/backend/resources/requirements.txt
anthropic==0.31.1
fastapi==0.111.0
langchain==0.2.14
langchain-anthropic==0.1.23
langchain-community==0.2.12
langchain-core==0.2.34
langchain-openai==0.1.16
numpy==1.26.4
openai==1.35.13
pandas==2.2.2
pydantic==2.8.2
pymongo==4.8.0
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
SQLAlchemy==2.0.31
sqlparse==0.5.1
starlette==0.37.2
tiktoken==0.7.0
typing_extensions==4.12.2
uvicorn==0.30.1
<<<<<<< HEAD:backend/resources/requirements.txt
psycopg2
=======
python-dotenv==1.0.1
tiktoken==0.7.0
>>>>>>> Shihab/FrontendUI:Backend/requirements.txt
